Break out File {Properties, Utilities}
